requirements.txt
python-docx
google_generativeai
streamlit